+++ src/direct_show/direct_show_camera.h
@@ -30,7 +30,7 @@ public:
     ~DirectShow_Camera();
     static std::unique_ptr<CameraInterface> createCamera(std::shared_ptr<void> mfDeinitializer, const CameraInformation &information);
 
-    int start(Format pixelFormat, int width, int height, int fps, frame_callback cb);
+    int start(Format pixelFormat, int width, int height, float fps, frame_callback cb);
     int stop();
     std::unique_ptr<PixelBuffer> CaptureFrame();
     // ---- Capabilities ----
@@ -44,7 +44,7 @@ private:
 
     /***** SDK FUNCTIONS *****/
     IMoniker* getIMonikerByUniqueId(std::shared_ptr<UniqueId> &uniqueId);    
-    int setCapabilities(ICaptureGraphBuilder2 *pBuild, IBaseFilter *pVCap, Format pixelFormat, int width, int height, int fps);
+    int setCapabilities(ICaptureGraphBuilder2 *pBuild, IBaseFilter *pVCap, Format pixelFormat, int width, int height, float fps);
 
 public:
     std::shared_ptr<void> mfDeinitializer;
