+++ src/media_foundation/media_foundation_camera.h
@@ -42,7 +42,7 @@ public:
     ~MediaFoundation_Camera();
     static std::unique_ptr<CameraInterface> createCamera(std::shared_ptr<void> mfDeinitializer, const CameraInformation &information);
 
-    int start(Format pixelFormat, int width, int height, int fps, frame_callback cb);
+    int start(Format pixelFormat, int width, int height, float fps, frame_callback cb);
     int stop();
     std::unique_ptr<PixelBuffer> CaptureFrame();  //TODO
     // ---- Capabilities ----
@@ -62,8 +62,8 @@ private:
     static int createVideoDeviceSource(const std::wstring &pszSymbolicLink, IMFMediaSource **ppSource);
     int getVideoCapabilities(IMFMediaSource *source, std::vector<CapabilityFormat> &capFormatVector) const;
     int setDeviceFormat(IMFMediaSource *source, const int width, const int height, const Format pixelFormat,
-                        const int fps) const;
-    int setReaderFormat(IMFSourceReader *reader, const int width, const int height, const Format pixelFormat, const int fps) const;
+                        const float fps) const;
+    int setReaderFormat(IMFSourceReader *reader, const int width, const int height, const Format pixelFormat, const float fps) const;
 
 public:
     std::shared_ptr<void> mfDeinitializer;
